The 'mocha' command is now used to run tests
Calling this using mocha for tests is a little bit of an exaggeration, but we should be able to do code coverage now.
Jeff, if you want to to switch things to properly use mocha, please feel free. Also if you want to include file names or test names in the output this would be great.