Fix https://github.com/WuTheFWasThat/vimflowy/issues/71 and a bug where cursors don't update when the block they're on is moved.
---
The root cause of this bug was that datastore has a cache, and instance objects are now complex enough that returned objects from the cache had shared memory (fixed with _.cloneDeep).

Therefore, getChildren was returning children that could affect one another -- the only instance where this mattered was when they were the SAME child under DIFFERENT parents, because 'getChildren' modifies the returned result to annotate the parent, which is not stored in the datastore.

This is something I had actually fixed once before I think, but it somehow got clobbered in the merge.

Cursors don't update
---
When indenting or dedenting, the cursor doesn't update, because previously it only stored the row id. Now since it points to an instance, it needs to store the whole ancestry, which means it needs to update if what it's pointing to moves. This bug was revealed by the previous bugfix.